+++ DfE.FindInformationAcademiesTrusts/Services/Academy/AcademyService.cs
@@ -6,6 +6,7 @@ public interface IAcademyService
 {
     Task<AcademyDetailsServiceModel[]> GetAcademiesInTrustDetailsAsync(string uid);
     Task<AcademyOfstedServiceModel[]> GetAcademiesInTrustOfstedAsync(string uid);
+    Task<AcademyPupilNumbersServiceModel[]> GetAcademiesInTrustPupilNumbersAsync(string uid);
 }
 
 public class AcademyService(IAcademyRepository academyRepository) : IAcademyService
@@ -26,5 +27,15 @@ public class AcademyService(IAcademyRepository academyRepository) : IAcademyServ
         return academies.Select(a =>
             new AcademyOfstedServiceModel(a.Urn, a.EstablishmentName, a.DateAcademyJoinedTrust, a.PreviousOfstedRating,
                 a.CurrentOfstedRating)).ToArray();
+    }
+
+    public async Task<AcademyPupilNumbersServiceModel[]> GetAcademiesInTrustPupilNumbersAsync(string uid)
+    {
+        var academies = await academyRepository.GetAcademiesInTrustPupilNumbersAsync(uid);
+
+        return academies.Select(a =>
+            new AcademyPupilNumbersServiceModel(a.Urn, a.EstablishmentName, a.PhaseOfEducation,
+                a.AgeRange, a.NumberOfPupils,
+                a.SchoolCapacity)).ToArray();
     }
 }