+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Changed
+
+- Reduced the load time of the Academies in trust pupil numbers page
+
+
 ## [Release-5][release-5] (production-2024-09-05.2971)
 
 ### Added